Create AVLTree.h
I have my AVL tree at the clean compile and seemingly bug free stage and so I am uploading all of my files to GitHub. Hopefully it will be a piece of cake when it comes to integrating it with the rest as its only external dependency is the Card class. I did expand the Card class to include overloads for all comparison operators so that they work by the cards code. I will be uploading 5 files in all and they are, AVLTree.h, AVLTree.cpp, AVLNode.h, LinkedList.h, and LinkedList.cpp.