Unifying all card output
I decided to take a stab at unifying all of the output. Since I am
changing other people's code and the format to be displayed isn't really
agreed upon I decided to create a new branch.

In this implementation all card searches and deletes display all of the
information about a card in (Code) (Name) (Cost) (Rarity) format. When
printing the entire hash table or BST the format (Code) (Name) is used.
and Finally when printing the AVL tree the format (Name) (Code1),
(Code2), ... is used.